Move HydeFront related methods to the new HydeFront facade
