+++ packages/framework/tests/Unit/Facades/HydeFrontFacadeTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Unit\Facades;
 
+use Hyde\Facades\HydeFront;
 use Hyde\Testing\UnitTestCase;
 
 /**
@@ -11,5 +12,14 @@ use Hyde\Testing\UnitTestCase;
  */
 class HydeFrontFacadeTest extends UnitTestCase
 {
-    //
+    public function testVersionReturnsString()
+    {
+        $this->assertIsString(HydeFront::version());
+    }
+
+    public function testCdnLinkReturnsCorrectUrl()
+    {
+        $expected = 'https://cdn.jsdelivr.net/npm/hydefront@v3.4/dist/styles.css';
+        $this->assertSame($expected, HydeFront::cdnLink('styles.css'));
+    }
 }
